Patterns: Fix pattern categories on import (#58926)
