+++ packages/patterns/src/components/create-pattern-modal.js
@@ -10,21 +10,17 @@ import {
 	ToggleControl,
 } from '@wordpress/components';
 import { __, _x } from '@wordpress/i18n';
-import { useState, useMemo } from '@wordpress/element';
-import { useDispatch, useSelect } from '@wordpress/data';
+import { useState } from '@wordpress/element';
+import { useDispatch } from '@wordpress/data';
 import { store as noticesStore } from '@wordpress/notices';
-import { store as coreStore } from '@wordpress/core-data';
 
 /**
  * Internal dependencies
  */
 import { PATTERN_DEFAULT_CATEGORY, PATTERN_SYNC_TYPES } from '../constants';
-
-/**
- * Internal dependencies
- */
 import { store as patternsStore } from '../store';
-import CategorySelector, { CATEGORY_SLUG } from './category-selector';
+import CategorySelector from './category-selector';
+import { useAddPatternCategory } from '../private-hooks';
 import { unlock } from '../lock-unlock';
 
 export default function CreatePatternModal( {
@@ -59,47 +55,9 @@ export function CreatePatternModalContents( {
 
 	const [ isSaving, setIsSaving ] = useState( false );
 	const { createPattern } = unlock( useDispatch( patternsStore ) );
-	const { saveEntityRecord, invalidateResolution } = useDispatch( coreStore );
 	const { createErrorNotice } = useDispatch( noticesStore );
 
-	const { corePatternCategories, userPatternCategories } = useSelect(
-		( select ) => {
-			const { getUserPatternCategories, getBlockPatternCategories } =
-				select( coreStore );
-
-			return {
-				corePatternCategories: getBlockPatternCategories(),
-				userPatternCategories: getUserPatternCategories(),
-			};
-		}
-	);
-
-	const categoryMap = useMemo( () => {
-		// Merge the user and core pattern categories and remove any duplicates.
-		const uniqueCategories = new Map();
-		userPatternCategories.forEach( ( category ) => {
-			uniqueCategories.set( category.label.toLowerCase(), {
-				label: category.label,
-				name: category.name,
-				id: category.id,
-			} );
-		} );
-
-		corePatternCategories.forEach( ( category ) => {
-			if (
-				! uniqueCategories.has( category.label.toLowerCase() ) &&
-				// There are two core categories with `Post` label so explicitly remove the one with
-				// the `query` slug to avoid any confusion.
-				category.name !== 'query'
-			) {
-				uniqueCategories.set( category.label.toLowerCase(), {
-					label: category.label,
-					name: category.name,
-				} );
-			}
-		} );
-		return uniqueCategories;
-	}, [ userPatternCategories, corePatternCategories ] );
+	const { categoryMap, findOrCreateTerm } = useAddPatternCategory();
 
 	async function onCreate( patternTitle, sync ) {
 		if ( ! title || isSaving ) {
@@ -137,38 +95,6 @@ export function CreatePatternModalContents( {
 		}
 	}
 
-	/**
-	 * @param {string} term
-	 * @return {Promise<number>} The pattern category id.
-	 */
-	async function findOrCreateTerm( term ) {
-		try {
-			const existingTerm = categoryMap.get( term.toLowerCase() );
-			if ( existingTerm && existingTerm.id ) {
-				return existingTerm.id;
-			}
-			// If we have an existing core category we need to match the new user category to the
-			// correct slug rather than autogenerating it to prevent duplicates, eg. the core `Headers`
-			// category uses the singular `header` as the slug.
-			const termData = existingTerm
-				? { name: existingTerm.label, slug: existingTerm.name }
-				: { name: term };
-			const newTerm = await saveEntityRecord(
-				'taxonomy',
-				CATEGORY_SLUG,
-				termData,
-				{ throwOnError: true }
-			);
-			invalidateResolution( 'getUserPatternCategories' );
-			return newTerm.id;
-		} catch ( error ) {
-			if ( error.code !== 'term_exists' ) {
-				throw error;
-			}
-
-			return error.data.term_id;
-		}
-	}
 	return (
 		<form
 			onSubmit={ ( event ) => {
